requirements.txt
madmom==0.16.dev0
matplotlib==3.0.0
pydub==0.23.0
networkx==2.2
pyacoustid==1.1.5
MDP==3.5
numpy==1.13.3
librosa==0.6.2
scipy==1.1.0
pandas
pyAudioAnalysis
eyeD3

